Elastsearchsmartlab/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
numpy>=1.24.0
openai>=1.0.0
cohere>=4.0.0
rank_bm25>=0.2.2
lunr>=0.6.2
requests>=2.31.0
google-generativeai>=0.3.0
